// packages/schedule/src/types.ts
export type CronOptions = {
  /**
   * Specify the name of your cron job. This will allow you to get reference to it via SchedulerRegistry.
   */
  name?: string;

  /**
   * Specify the timezone for the execution.
   */
  timeZone?: string;

  /**
   * This allows you to specify the offset of your timezone rather than using the timeZone param.
   */
  utcOffset?: number;

  /**
   * If you have code that keeps the event loop running and want to stop the node process when that finishes, you can do so.
   */
  unrefTimeout?: boolean;

  /**
   * If true, no additional instances of cronjob will run until the current onTick callback has completed.
   */
  waitForCompletion?: boolean;

  /**
   * This flag indicates whether the job will be executed at all.
   * @default false
   */
  disabled?: boolean;

  /**
   * Threshold in ms to control whether to execute or skip missed execution deadlines caused by slow or busy hardware.
   * Default is 250
   */
  threshold?: number;

  /**
   * Delay in milliseconds before the first cron execution after application bootstrap.
   */
  initialDelay?: number;
};